+++ src/common/configuration.cpp
@@ -96,6 +96,12 @@ Configuration::Configuration(const fs::path& config_path,
                 } else {
                     invalid_options.push_back(key);
                 }
+            } else if (key == "frame_rate") {
+                if (const auto parsed_value = value.as_floating_point()) {
+                    frame_rate = parsed_value->get();
+                } else {
+                    invalid_options.push_back(key);
+                }
             } else if (key == "group") {
                 if (const auto parsed_value = value.as_string()) {
                     group = parsed_value->get();
@@ -109,4 +115,9 @@ Configuration::Configuration(const fs::path& config_path,
 
         break;
     }
+}
+
+std::chrono::steady_clock::duration Configuration::event_loop_interval() const {
+    return std::chrono::duration_cast<std::chrono::steady_clock::duration>(
+        std::chrono::milliseconds(1000) / frame_rate.value_or(60.0));
 }